Update eip-0021.md
Proposed change: adding GNOMECOIN to the list of verified tokens
Reasoning: issued in March 2021, GNOMECOIN is the first token issued on the Ergo blockchain to have utility in the NFT space, letting holders exchange them for special or custom NFTs. As more benfits are added for GNOMECOIN holders, preventing scams by having the token verified on wallets would be extremely useful.